Don't include PureJSON in Go 1.6 and lower
